Refactor migration and todo index
The migration will now drop the `todos` table and repopulate it each
time the app is started. This ibviously isn't a good long-term solution,
but during development it is easier than dropping the table or changing
the seed data manually.
The table is now seeded with four rows including both completed and
non-completed todos. The index page pulls the data in order by completed
state (non-completed first, of course) and then in the order in which
they were added to the database.
The index page of the app now shows a simple HTML table with id, body,
and completed state which makes it easier to see that the sorting is
working.