Refactor file structure
`Handler` was getting pretty crowded and it was confusing me. Now the
logic is pulled out into:

* Views - Handles presentation concerns
* Model - Handles data access
* Router - Routes HTTP requests to the correct place

I chose to skip the idea of something like a controller here because I'm
not sure that the complexity of this app warrants it. That wouls
probably be something to think about in the future.

Views may have to be broken out a little more in the future as well.

I also renamed the handle "App" because it doesn't handle too much else
right now.

Another reason to do all of this was to learn a little more about files
and `ns`.